Rename `Align to corner of the screen` setting to `Align to edge of taskbar`
The setting was improperly named and evoked that custom button should be
"moved" to the corner of screen even for center aligned taskbar on Windows 11.

What the setting does (and always did) is to align button to the edge of
taskbar (e.g. bottom edge in case of bottom taskbar) instead of centering it.

Hopefully it will be now more clear.